move video.py trimming 8MB
