+++ instagram_to_discord/__init__.py
@@ -9,6 +9,7 @@ from . import cookie_requests
 from . import download
 from . import youtube
 from . import tiktok
+from . import video
 import os
